Move cuda out of src directory and use fully qualified #includes.
