+++ cuda/ArrayUtils.cpp
