+++ cuda/DeviceOpaqueArray2D.inl
