+++ cuda/KernelArray1D.inl
