+++ cuda/Rect2f.inl
@@ -1,4 +1,6 @@
-libcgt::cuda::Rect2f::Rect2f() :
+namespace libcgt { namespace cuda {
+
+Rect2f::Rect2f() :
 
     m_origin( make_float2( 0.f, 0.f ) ),
     m_size( make_float2( 0.f, 0.f ) )
@@ -7,7 +9,7 @@ libcgt::cuda::Rect2f::Rect2f() :
 
 }
 
-libcgt::cuda::Rect2f::Rect2f( float left, float bottom, float width, float height ) :
+Rect2f::Rect2f( float left, float bottom, float width, float height ) :
 
     m_origin( make_float2( left, bottom ) ),
     m_size( make_float2( width, height ) )
@@ -16,7 +18,7 @@ libcgt::cuda::Rect2f::Rect2f( float left, float bottom, float width, float heigh
 
 }
 
-libcgt::cuda::Rect2f::Rect2f( float width, float height ) :
+Rect2f::Rect2f( float width, float height ) :
 
     m_origin( make_float2( 0.f, 0.f ) ),
     m_size( make_float2( width, height ) )
@@ -25,7 +27,7 @@ libcgt::cuda::Rect2f::Rect2f( float width, float height ) :
 
 }
 
-libcgt::cuda::Rect2f::Rect2f( const float2& origin, const float2& size ) :
+Rect2f::Rect2f( const float2& origin, const float2& size ) :
 
     m_origin( origin ),
     m_size( size )
@@ -34,7 +36,7 @@ libcgt::cuda::Rect2f::Rect2f( const float2& origin, const float2& size ) :
 
 }
 
-libcgt::cuda::Rect2f::Rect2f( const float2& size ) :
+Rect2f::Rect2f( const float2& size ) :
 
     m_origin( make_float2( 0.f, 0.f ) ),
     m_size( size )
@@ -43,65 +45,65 @@ libcgt::cuda::Rect2f::Rect2f( const float2& size ) :
 
 }
 
-float libcgt::cuda::Rect2f::left() const
+float Rect2f::left() const
 {
     return m_origin.x;
 }
 
-float libcgt::cuda::Rect2f::right() const
+float Rect2f::right() const
 {
     return m_origin.x + m_size.x;
 }
 
-float libcgt::cuda::Rect2f::bottom() const
+float Rect2f::bottom() const
 {
     return m_origin.y;
 }
 
-float libcgt::cuda::Rect2f::top() const
+float Rect2f::top() const
 {
     return m_origin.y + m_size.y;
 }
 
-float2 libcgt::cuda::Rect2f::bottomLeft() const
+float2 Rect2f::bottomLeft() const
 {
     return m_origin;
 }
 
-float2 libcgt::cuda::Rect2f::bottomRight() const
+float2 Rect2f::bottomRight() const
 {
     return make_float2( m_origin.x + m_size.x, m_origin.y );
 }
 
-float2 libcgt::cuda::Rect2f::topLeft() const
+float2 Rect2f::topLeft() const
 {
     return make_float2( m_origin.x, m_origin.y + m_size.y );
 }
 
-float2 libcgt::cuda::Rect2f::topRight() const
+float2 Rect2f::topRight() const
 {
     return m_origin + m_size;
 }
 
-float2 libcgt::cuda::Rect2f::origin() const
+float2 Rect2f::origin() const
 {
     return m_origin;
 }
 
-float2 libcgt::cuda::Rect2f::size() const
+float2 Rect2f::size() const
 {
     return m_size;
 }
 
-float libcgt::cuda::Rect2f::area() const
+float Rect2f::area() const
 {
     return m_size.x * m_size.y;
 }
 
-libcgt::cuda::Rect2i libcgt::cuda::Rect2f::enlargedToInt() const
+Rect2i Rect2f::enlargedToInt() const
 {
-    int2 minimum = libcgt::cuda::floorToInt( bottomLeft() );
-    int2 maximum = libcgt::cuda::floorToInt( topRight() );
+    int2 minimum = floorToInt( bottomLeft() );
+    int2 maximum = floorToInt( topRight() );
 
     // size does not need a +1:
     // say min is 1.1 and max is 3.6
@@ -109,18 +111,18 @@ libcgt::cuda::Rect2i libcgt::cuda::Rect2f::enlargedToInt() const
     // hence, we want indices 1, 2, 3, which has size 3 = 4 - 1
     int2 size = maximum - minimum;
 
-    return libcgt::cuda::Rect2i( minimum, size );
+    return Rect2i( minimum, size );
 }
 
 // static
-bool libcgt::cuda::Rect2f::intersect( const libcgt::cuda::Rect2f& r0, const libcgt::cuda::Rect2f& r1 )
+bool Rect2f::intersect( const Rect2f& r0, const Rect2f& r1 )
 {
-    libcgt::cuda::Rect2f isect;
+    Rect2f isect;
     return intersect( r0, r1, isect );
 }
 
 // static
-bool libcgt::cuda::Rect2f::intersect( const libcgt::cuda::Rect2f& r0, const libcgt::cuda::Rect2f& r1, libcgt::cuda::Rect2f& intersection )
+bool Rect2f::intersect( const Rect2f& r0, const Rect2f& r1, Rect2f& intersection )
 {
     float2 minimum = fmaxf( r0.bottomLeft(), r1.bottomLeft() );
     float2 maximum = fminf( r0.topRight(), r1.topRight() );
@@ -134,3 +136,5 @@ bool libcgt::cuda::Rect2f::intersect( const libcgt::cuda::Rect2f& r0, const libc
     }
     return false;
 }
+
+} } // cuda, libcgt
